refactor: 새로고침 토큰 추가 로직 분리
